requirements.txt
python-dotenv==1.0.1
openai==1.40.0
streamlit==1.37.0
requests==2.32.3
beautifulsoup4==4.12.3
